pyproject.toml
[project]
name = "lilt"
version = "0.1.0"
description = "Local-first stack to chat with your documents, the agentic way."
authors = [
    {name = "Duy Nguyen", email = "57234880+duynguyen158@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">= 3.13"
dependencies = [
    "python-documentcloud>=4.5.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
]

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "F", # Pyflakes
    "UP", # pyupgrade
    "B", # flake8-bugbear
    "SIM", # flake8-simplify
    "I", # isort
]
ignore = [
    "E501", # line too long
]

[tool.ruff.lint.isort]
# If specifying known third-party packages is useful, e.g., so that isort works correctly, uncomment and fill in the list below.
# known-third-party = []
combine-as-imports = true

[tool.ty]
# Environment settings
[tool.ty.environment]
root = ["./lilt", "./tests", "./scripts"]

[tool.ty.analysis]
allowed-unresolved-imports = []
replace-imports-with-any = []

[tool.ty.src]
include = ["lilt", "tests", "scripts"]
respect-ignore-files = true
